Role class added and User class Refactored from business package to user package
