+++ src/main/java/com/dev02/libraryproject/entity/concretes/user/Role.java
@@ -1,0 +1,29 @@
+package com.dev02.libraryproject.entity.concretes.user;
+
+import com.dev02.libraryproject.entity.enums.RoleType;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+
+@Entity
+@Table(name = "roles")
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Role {
+    @Id
+    @GeneratedValue(strategy =  GenerationType.IDENTITY)
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(length = 20)
+    @NotNull
+    private RoleType name;
+
+}
